+++ riemann/divisor_test.go
@@ -14,7 +14,7 @@ var _ = Describe("Divisor", func() {
 			if err != nil {
 				Fail("error should be nil")
 			}
-			Expect(output).To(Equal(195))
+			Expect(output).To(Equal(int64(195)))
 		})
 
 		It("Should give correct result", func() {
@@ -22,7 +22,7 @@ var _ = Describe("Divisor", func() {
 			if err != nil {
 				Fail("error should be nil")
 			}
-			Expect(output).To(Equal(1))
+			Expect(output).To(Equal(int64(1)))
 		})
 	})
 
@@ -41,7 +41,7 @@ var _ = Describe("Divisor", func() {
 	})
 
 	It("Should work on parametrized cases", func() {
-		divisorSums := []int{
+		divisorSums := []int64{
 			1, 3, 4, 7, 6, 12, 8, 15, 13, 18, 12, 28, 14, 24, 24, 31, 18, 39, 20, 42, 32,
 			36, 24, 60, 31, 42, 40, 56, 30, 72, 32, 63, 48, 54, 48, 91, 38, 60, 56, 90, 42,
 			96, 44, 84, 78, 72, 48, 124, 57, 93, 72,
@@ -49,7 +49,7 @@ var _ = Describe("Divisor", func() {
 
 		for i, sum := range divisorSums {
 			By("Passing each cases", func() {
-				output, err := riemann.DivisorSum(i + 1)
+				output, err := riemann.DivisorSum(int64(i) + 1)
 				if err != nil {
 					Fail("error should be nil")
 				}
